+++ pyinaturalist/controllers/__init__.py
@@ -2,4 +2,5 @@
 # flake8: noqa: F401
 from pyinaturalist.controllers.base import BaseController
 from pyinaturalist.controllers.observations import ObservationController
+from pyinaturalist.controllers.projects import ProjectController
 from pyinaturalist.controllers.taxa import TaxonController