+++ test/v1/test_projects.py
@@ -55,7 +55,7 @@ def test_add_project_observation(requests_mock):
         json=load_sample_data('add_project_observation.json'),
         status_code=200,
     )
-    response = add_project_observation('token', project_id=1234, observation_id=5678)
+    response = add_project_observation(project_id=1234, observation_id=5678, access_token='token')
     assert response['id'] == 54986584
 
 
@@ -64,5 +64,5 @@ def test_delete_project_observation(requests_mock):
         f'{API_V1_BASE_URL}/projects/1234/remove',
         status_code=200,
     )
-    response = delete_project_observation('token', project_id=1234, observation_id=5678)
+    response = delete_project_observation(project_id=1234, observation_id=5678, access_token='token')
     assert response.status_code == 200