+++ HISTORY.md
@@ -25,6 +25,7 @@
 ### Other Changes
 * Allow all API request functions to accept a `limiter` argument to override rate-limiting settings
 * Allow all API request functions to accept a `dry_run` argument to dry-run an individual request
+* Fix bug with `rule_details` param for `get_projects_by_id()`
 
 ## 0.14.1 (2021-07-21)
 * Added new function for **Posts** endpoint: `get_posts()`